SoC: Fix frequency print gcc warning
type 'unsigned int', but argument 2 has type 'long unsigned int'

+++ SoC/demosoc/Common/Source/system_demosoc.c
@@ -286,7 +286,7 @@ void SystemBannerPrint(void)
 #ifdef DOWNLOAD_MODE_STRING
     printf("Download Mode: %s\r\n", DOWNLOAD_MODE_STRING);
 #endif
-    printf("CPU Frequency %lu Hz\r\n", SystemCoreClock);
+    printf("CPU Frequency %u Hz\r\n", (unsigned int)SystemCoreClock);
 #endif
 }
